[Common UI] Generalize click-away behavior
Add a general-purpose controller for UI elements which
have 'click-away' behavior; that is, they should be
deactivated on document clicks.

This generalizes existing behavior added for the Create
menu, such that it may be used on other, similar menus
and UI elements.

Part of ongoing transition of common user interface
components, WTD-574.