[Persistence] Move persistence sources
Move persistence sources into a deeper bundle in preparation for
splitting of persistence into two bundles - one specific to the
Couch adapter, and one for persistence caching generally. This
supports WTD-702, the WARP persistence adapter, which will not
use Couch for persistence but which will want to use the platform-
provided persistence cache.