[Tree] Check for change before scope.$apply
TreeView's observers will be called when the selected domain object
changes, which can occur for one of two reasons:

1. Because a new value was set externally, from mct-tree.
2. Because a new value was selected, by the user.

In the latter case a $apply is needed, but in the former it is not
(and causes an error.) However, when that case occurs, the value in
scope will be up to date already (it was a watch that triggered the
call to treeView.value) so no assignment or $apply is necessary.

Fixes #1114.